backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=0.21
# Use the "srv" extra to ensure DNS SRV support (dnspython) for MongoDB Atlas
pymongo[srv]>=4.0
openai>=2.0
